+++ vscode/pacioli/CHANGELOG.md
@@ -2,8 +2,11 @@
 
 All notable changes to the "pacioli" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## 0.0.0
 
-## [Unreleased]
+Initial release
 
-- Initial release
+### Added
+
+- Syntax highlighting
+- Diagnostics for syntax errors